+++ src/app/components/listing_details/criteria.directive.js
@@ -15,23 +15,15 @@
                 cert: '=',
                 hasIcs: '=',
                 viewAll: '=',
+                hasEdited: '&',
                 isEditing: '=',
                 resources: '=',
                 accessibilityStandards: '=?',
                 qmsStandards: '=?',
-                save: '&',
             },
             scope: {},
             controllerAs: 'vm',
             controller: 'CertificationCriteriaController',
-            link: function (scope, element, attr, ctrl) {
-                var handler = ctrl.save({
-                    handler: function () {
-                        ctrl.saveEdits();
-                    },
-                });
-                scope.$on('$destroy', handler);
-            },
         };
     }
 
@@ -40,7 +32,7 @@
         var vm = this;
 
         vm.editCert = editCert;
-        vm.saveEdits = saveEdits;
+        vm.hasPhantomData = hasPhantomData;
         vm.toggleCriteria = toggleCriteria;
 
         activate();
@@ -66,6 +58,7 @@
                 },
             });
             vm.editUibModalInstance.result.then(function (result) {
+                vm.hasEdited();
                 vm.cert = result;
             }, function (result) {
                 if (result !== 'cancelled') {
@@ -74,22 +67,24 @@
             });
         }
 
-        /*
-         * remove any keys where the select used 'vm.options' and the display was N/A
-         */
-        function saveEdits () {
-            if (vm.cert.gap === 'null') {
-                delete (vm.cert.gap);
-            }
-            if (vm.cert.g1Success === 'null') {
-                delete (vm.cert.g1Success);
-            }
-            if (vm.cert.g2Success === 'null') {
-                delete (vm.cert.g2Success);
-            }
-            if (vm.cert.sed === 'null') {
-                delete (vm.cert.sed);
-            }
+        function hasPhantomData () {
+            var ret =
+                (vm.cert.additionalSoftware && vm.cert.additionalSoftware.length > 0) ||
+                (vm.cert.apiDocumentation && vm.cert.apiDocumentation.length > 0) ||
+                (vm.cert.g1MacraMeasures && vm.cert.g1MacraMeasures.length > 0) ||
+                (vm.cert.g1Success) ||
+                (vm.cert.g2MacraMeasures && vm.cert.g2MacraMeasures.length > 0) ||
+                (vm.cert.g2Success) ||
+                (vm.cert.gap) ||
+                (vm.cert.privacySecurityFramework && vm.cert.privacySecurityFramework.length > 0) ||
+                (vm.cert.sed) ||
+                (vm.cert.testDataUsed && vm.cert.testDataUsed.length > 0) ||
+                (vm.cert.testFunctionality && vm.cert.testFunctionality.length > 0) ||
+                (vm.cert.testProcedures && vm.cert.testProcedures.length > 0) ||
+                (vm.cert.testStandards && vm.cert.testStandards.length > 0) ||
+                (vm.cert.testToolsUsed && vm.cert.testToolsUsed.length > 0) ||
+                false;
+            return ret;
         }
 
         function toggleCriteria () {