+++ src/app/components/listing_details/details.directive.js
@@ -41,7 +41,9 @@
             var vm = this;
 
             vm.ACTIVE_CAP = ACTIVE_CAP;
+            vm.hasEdited = hasEdited;
             vm.prepCqms = prepCqms
+            vm.registerSed = registerSed;
             vm.saveEdits = saveEdits;
             vm.sortCerts = sortCerts;
             vm.sortCqms = sortCqms;
@@ -53,6 +55,7 @@
             ////////////////////////////////////////////////////////////////////
 
             function activate () {
+                vm.handlers = [];
                 if (angular.isUndefined(vm.isEditing)) {
                     vm.isEditing = false;
                 }
@@ -77,6 +80,12 @@
                     }}, true);
             }
 
+            function hasEdited () {
+                angular.forEach(vm.handlers, function (handler) {
+                    handler();
+                });
+            }
+
             function prepCqms () {
                 if (vm.cqms) {
                     for (var i = 0; i < vm.cqms.length; i++) {
@@ -86,6 +95,16 @@
                         }
                     }
                 }
+            }
+
+            function registerSed (handler) {
+                vm.handlers.push(handler);
+                var removeHandler = function () {
+                    vm.handlers = vm.handlers.filter(function (aHandler) {
+                        return aHandler !== handler;
+                    });
+                };
+                return removeHandler;
             }
 
             function saveEdits () {